requirements.txt
asgiref==3.8.1
coloredlogs==15.0.1
distlib==0.3.9
Django==5.1.6
django-codemirror6==1.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.18.0
humanfriendly==10.0
pillow==11.2.1
platformdirs==4.3.8
PyJWT==2.10.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-printr==4.9
python-timeout==6.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.31.2
google-cloud-aiplatform>=1.38.0
